requirements.txt
# Stack: Flask + Gunicorn + Celery + PostgreSQL + Redis
Flask>=3.0.0
gunicorn>=21.0.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dateutil>=2.8.0
odfpy>=1.4.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.3
bcrypt>=4.0.0
PyJWT>=2.8.0